Update sqr_clipped_relu.h
Bench: 2405297

+++ src/nnue/layers/sqr_clipped_relu.h
@@ -101,8 +101,7 @@ class SqrClippedReLU final {
             output[i] = static_cast<OutputType>(
               // Really should be /127 but need to make it fast so right-shift
               // by an extra 7 bits instead. Needs to be accounted for in the trainer.
-              std::min(((long long) (input[i]) * input[i]) >> (7 + 2 * WEIGHT_SCALE_BITS),
-                       i64{127}));
+              std::min((i64(input[i]) * input[i]) >> (7 + 2 * WEIGHT_SCALE_BITS), i64{127}));
         }
     }
 };
